Refactor, clean code and remove old Coin Manager
