+++ app/src/main/AndroidManifest.xml
@@ -19,8 +19,6 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.AppTheme.DayNight"
         tools:ignore="GoogleAppIndexingWarning">
-        <activity android:name=".modules.managewallets.ManageWalletsActivityMvvm"/>
-
         <meta-data
             android:name="com.google.firebase.messaging.default_notification_icon"
             android:resource="@drawable/ic_notification_icon" />
@@ -66,7 +64,13 @@
             android:name=".modules.lockscreen.LockScreenActivity"
             android:screenOrientation="portrait" />
         <activity
+            android:name="io.horizontalsystems.pin.PinActivity"
+            android:screenOrientation="portrait" />
+        <activity
             android:name=".modules.managecoins.views.ManageWalletsActivity"
+            android:screenOrientation="portrait" />
+        <activity
+            android:name=".modules.settings.managekeys.views.ManageKeysActivity"
             android:screenOrientation="portrait" />
         <activity
             android:name=".modules.restore.RestoreActivity"
