+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/managewallets/view/ManageWalletsFragment.kt
@@ -1,4 +1,4 @@
-package io.horizontalsystems.bankwallet.modules.managewallets.ui.main
+package io.horizontalsystems.bankwallet.modules.managewallets.view
 
 import android.content.Intent
 import android.os.Bundle
@@ -7,12 +7,16 @@ import android.view.View
 import android.view.ViewGroup
 import android.view.inputmethod.InputMethodManager
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentActivity
+import androidx.fragment.app.commit
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import io.horizontalsystems.bankwallet.R
+import io.horizontalsystems.bankwallet.core.BaseFragment
 import io.horizontalsystems.bankwallet.core.putParcelableExtra
 import io.horizontalsystems.bankwallet.core.utils.ModuleField
 import io.horizontalsystems.bankwallet.entities.*
+import io.horizontalsystems.bankwallet.modules.addErc20token.AddErc20TokenActivity
 import io.horizontalsystems.bankwallet.modules.managewallets.ManageWalletsModule
 import io.horizontalsystems.bankwallet.modules.noaccount.NoAccountDialog
 import io.horizontalsystems.bankwallet.modules.restore.RestoreActivity
@@ -22,10 +26,15 @@ import io.horizontalsystems.bankwallet.ui.helpers.AppLayoutHelper
 import io.horizontalsystems.views.TopMenuItem
 import kotlinx.android.synthetic.main.manage_wallets_fragment.*
 
-class ManageWalletsFragment : Fragment(), ManageWalletItemsAdapter.Listener, NoAccountDialog.Listener {
+class ManageWalletsFragment : BaseFragment(), ManageWalletItemsAdapter.Listener, NoAccountDialog.Listener {
 
     companion object {
-        fun newInstance() = ManageWalletsFragment()
+        fun start(activity: FragmentActivity) {
+            activity.supportFragmentManager.commit {
+                add(R.id.fragmentContainerView, ManageWalletsFragment())
+                addToBackStack(null)
+            }
+        }
     }
 
     private lateinit var viewModel: ManageWalletsViewModel
@@ -40,8 +49,8 @@ class ManageWalletsFragment : Fragment(), ManageWalletItemsAdapter.Listener, NoA
         super.onViewCreated(view, savedInstanceState)
         shadowlessToolbar.bind(
                 getString(R.string.ManageCoins_title),
-                leftBtnItem = TopMenuItem(text = R.string.ManageCoins_AddToken, onClick = { }),
-                rightBtnItem = TopMenuItem(text = R.string.Button_Done, onClick = { })
+                leftBtnItem = TopMenuItem(text = R.string.ManageCoins_AddToken, onClick = { showAddTokenDialog() }),
+                rightBtnItem = TopMenuItem(text = R.string.Button_Done, onClick = { activity?.onBackPressed() })
         )
 
         viewModel = ViewModelProvider(this, ManageWalletsModule.Factory())
@@ -95,9 +104,7 @@ class ManageWalletsFragment : Fragment(), ManageWalletItemsAdapter.Listener, NoA
         })
 
         viewModel.openDerivationSettingsLiveEvent.observe(viewLifecycleOwner, Observer { (coin, currentDerivation) ->
-
-            activity?.getSystemService(InputMethodManager::class.java)?.hideSoftInputFromWindow(activity?.currentFocus?.windowToken, 0)
-
+            hideKeyboard()
             showAddressFormatSelectionDialog(coin, currentDerivation)
         })
     }
@@ -122,5 +129,20 @@ class ManageWalletsFragment : Fragment(), ManageWalletItemsAdapter.Listener, NoA
                     viewModel.onCancelDerivationSelection()
                 }
         )
+    }
+
+    private fun showAddTokenDialog() {
+        hideKeyboard()
+        activity?.let {
+            AddTokenDialog.show(it, object : AddTokenDialog.Listener {
+                override fun onClickAddErc20Token() {
+                    startActivity(Intent(activity, AddErc20TokenActivity::class.java))
+                }
+            })
+        }
+    }
+
+    private fun hideKeyboard() {
+        activity?.getSystemService(InputMethodManager::class.java)?.hideSoftInputFromWindow(activity?.currentFocus?.windowToken, 0)
     }
 }
