+++ app/src/main/res/layout/manage_wallets_fragment.xml
@@ -5,6 +5,8 @@
     android:id="@+id/manage_wallets"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:background="?android:windowBackground"
+    android:clickable="true"
     app:layoutDescription="@xml/manage_wallets_fragment_scene">
 
     <io.horizontalsystems.views.ShadowlessToolbarView
@@ -19,8 +21,8 @@
         android:id="@+id/searchView"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:layout_marginTop="10dp"
         android:layout_marginStart="16dp"
+        android:layout_marginTop="10dp"
         android:layout_marginEnd="16dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
